OPENNLP-1894: Address lattice tokenizer review feedback
