+++ opennlp-core/opennlp-runtime/src/main/java/opennlp/tools/tokenize/lattice/MecabDictionaryInstaller.java
@@ -24,19 +24,16 @@ import java.net.URI;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.StandardCopyOption;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.zip.GZIPInputStream;
 
-import opennlp.tools.util.DictionaryCatalog;
-import opennlp.tools.util.DownloadUtil;
+import opennlp.tools.util.ResourceLimits;
 import opennlp.tools.util.model.UncloseableInputStream;
 
 /**
- * Fetches and unpacks a MeCab-format dictionary archive into a local directory, so the
- * dictionary is acquired by the user at install time and never ships with this library.
- * No dictionary data is bundled. Any non-{@code file:} archive is downloaded through
- * {@link DownloadUtil#download(URI, Path, String)} and requires an expected SHA-512
- * digest. Built-in catalog URLs are opt-in via {@link #installFromCatalog(String, Path)}.
+ * Unpacks a local MeCab-format dictionary archive into a directory. No dictionary data
+ * is bundled with this library.
  *
  * <p>The installer reads gzip-compressed
  * <a href="https://pubs.opengroup.org/onlinepubs/9699919799/utilities/pax.html#tag_20_92_13_06">
@@ -47,14 +44,14 @@ import opennlp.tools.util.model.UncloseableInputStream;
  * {@code *.def} definition files that a {@link MecabDictionary} reads, plus the
  * {@code dicrc} configuration file distributions ship alongside them, taken from the
  * archive root only (at most one leading directory deep). Deeper entries are skipped:
- * mecab-ko-dic, for example, nests {@code user-dic} templates whose numeric fields are
- * empty because they are input for {@code mecab-dict-index}, not loadable lexicon
- * data. Extracted entries are flattened to their base names, which also means no
+ * mecab-ko-dic, for example, nests {@code user-dic} templates with empty numeric fields
+ * because they are input for {@code mecab-dict-index}, not loadable lexicon data.
+ * Extracted entries are flattened to their base names, which also means no
  * archive path can escape the target directory.</p>
  *
  * <p>Extraction is bounded: each entry's declared size, the total bytes written, the
  * number of extracted dictionary files, and the gzip expansion ratio each have an
- * explicit ceiling so a crafted archive cannot fill the disk. The byte ceilings can be
+ * explicit limit so a crafted archive cannot fill the disk. The byte limits can be
  * raised at JVM startup via {@link #MAX_ENTRY_BYTES_PROPERTY} and
  * {@link #MAX_TOTAL_EXTRACTED_BYTES_PROPERTY} for dictionaries larger than the
  * defaults, such as UniDic.</p>
@@ -67,15 +64,20 @@ public final class MecabDictionaryInstaller {
   private static final int TAR_NAME_LENGTH = 100;
   private static final int TAR_SIZE_OFFSET = 124;
   private static final int TAR_SIZE_LENGTH = 12;
+  private static final int TAR_CHECKSUM_OFFSET = 148;
+  private static final int TAR_CHECKSUM_LENGTH = 8;
   private static final int TAR_TYPE_OFFSET = 156;
+  private static final byte TAR_CHECKSUM_SPACE = ' ';
+  private static final String STAGING_PREFIX = ".mecab-staging-";
 
   /**
    * System property for overriding {@link #MAX_ENTRY_BYTES}. Set at JVM startup,
    * e.g. {@code -Dopennlp.install.max.entry.bytes=2147483648} for dictionaries whose
-   * lexicon files exceed the default ceiling. Falls back to the default if absent,
+   * lexicon files exceed the default limit. Falls back to the default if absent,
    * non-numeric, or not positive.
    */
-  public static final String MAX_ENTRY_BYTES_PROPERTY = "opennlp.install.max.entry.bytes";
+  public static final String MAX_ENTRY_BYTES_PROPERTY =
+      ResourceLimits.MAX_ARCHIVE_ENTRY_BYTES_PROPERTY;
 
   /**
    * System property for overriding {@link #MAX_TOTAL_EXTRACTED_BYTES}. Set at JVM
@@ -83,39 +85,35 @@ public final class MecabDictionaryInstaller {
    * the default if absent, non-numeric, or not positive.
    */
   public static final String MAX_TOTAL_EXTRACTED_BYTES_PROPERTY =
-      "opennlp.install.max.total.bytes";
+      ResourceLimits.MAX_ARCHIVE_TOTAL_BYTES_PROPERTY;
 
   /**
-   * Inclusive ceiling on one tar entry's declared size, in bytes: 512 MiB unless
+   * Inclusive limit on one tar entry's declared size, in bytes: 512 MiB unless
    * overridden via {@link #MAX_ENTRY_BYTES_PROPERTY}.
    */
-  static final long MAX_ENTRY_BYTES =
-      DownloadUtil.configuredLimit(MAX_ENTRY_BYTES_PROPERTY, 512L * 1024 * 1024);
+  static final long MAX_ENTRY_BYTES = ResourceLimits.MAX_ARCHIVE_ENTRY_BYTES;
 
   /**
-   * Inclusive ceiling on the sum of extracted dictionary file sizes, in bytes: 2 GiB
+   * Inclusive limit on the sum of extracted dictionary file sizes, in bytes: 2 GiB
    * unless overridden via {@link #MAX_TOTAL_EXTRACTED_BYTES_PROPERTY}.
    */
-  static final long MAX_TOTAL_EXTRACTED_BYTES =
-      DownloadUtil.configuredLimit(MAX_TOTAL_EXTRACTED_BYTES_PROPERTY,
-          2L * 1024 * 1024 * 1024);
+  static final long MAX_TOTAL_EXTRACTED_BYTES = ResourceLimits.MAX_ARCHIVE_TOTAL_BYTES;
 
-  /** Inclusive ceiling on the number of dictionary files extracted from one archive. */
+  /** Inclusive limit on the number of dictionary files extracted from one archive. */
   static final int MAX_EXTRACTED_ENTRIES = 10_000;
 
   /**
-   * Inclusive ceiling on decompressed bytes per compressed byte while reading the
+   * Inclusive limit on decompressed bytes per compressed byte while reading the
    * gzip wrapper; higher expansion fails before the payload is written.
    */
   static final int MAX_GZIP_EXPANSION_RATIO = 100;
 
   private MecabDictionaryInstaller() {
-    // This class exposes only static methods and is never instantiated.
+    // This class exposes only static methods and cannot be instantiated.
   }
 
   /**
-   * Unpacks a local {@code file:} archive URI. Any other scheme requires
-   * {@link #install(URI, Path, String)} with an expected SHA-512 digest.
+   * Unpacks a trusted local {@code file:} archive URI.
    *
    * @param archive The archive location, a gzip-compressed ustar tar. Must not be
    *                {@code null}.
@@ -123,35 +121,13 @@ public final class MecabDictionaryInstaller {
    *                        be {@code null}.
    * @return The number of dictionary files extracted.
    * @throws IOException Thrown if reading or writing fails, the archive contains no
-   *         dictionary file, or an extraction budget is exceeded.
+   *         dictionary file, an extraction budget is exceeded, or the target already
+   *         contains a dictionary file with the same name.
    * @throws IllegalArgumentException Thrown if a parameter is {@code null},
    *         {@code archive} is not an absolute URI, or {@code archive} is not a
    *         {@code file:} URI.
    */
   public static int install(URI archive, Path targetDirectory) throws IOException {
-    return install(archive, targetDirectory, null);
-  }
-
-  /**
-   * Downloads a dictionary archive when needed, verifies its SHA-512 digest through
-   * {@link DownloadUtil#download(URI, Path, String)}, and unpacks it. A {@code file:}
-   * URI may omit the digest and is then opened without verification.
-   *
-   * @param archive The archive location, a gzip-compressed ustar tar. Must not be
-   *                {@code null}.
-   * @param targetDirectory The directory to unpack into; created when absent. Must not
-   *                        be {@code null}.
-   * @param expectedSha512 The expected SHA-512 hex digest. Required for any
-   *                       non-{@code file:} URI; optional for {@code file:} URIs.
-   * @return The number of dictionary files extracted.
-   * @throws IOException Thrown if fetching, verification, reading, or writing fails,
-   *         the archive contains no dictionary file, or an extraction budget is exceeded.
-   * @throws IllegalArgumentException Thrown if a parameter is {@code null},
-   *         {@code archive} is not an absolute URI, or a non-{@code file:} URI omits
-   *         the digest.
-   */
-  public static int install(URI archive, Path targetDirectory, String expectedSha512)
-      throws IOException {
     if (archive == null) {
       throw new IllegalArgumentException("archive must not be null");
     }
@@ -161,63 +137,12 @@ public final class MecabDictionaryInstaller {
     if (!archive.isAbsolute()) {
       throw new IllegalArgumentException("archive must be an absolute URI");
     }
-    if (expectedSha512 == null) {
-      if (!isLocalFile(archive)) {
-        throw new IllegalArgumentException("a non-file archive requires an expected "
-            + "SHA-512 digest; use install(URI, Path, String)");
-      }
-      try (InputStream in = archive.toURL().openStream()) {
-        return extract(in, targetDirectory);
-      }
+    if (!"file".equalsIgnoreCase(archive.getScheme())) {
+      throw new IllegalArgumentException("archive must use the file scheme");
     }
-    final Path downloaded = Files.createTempFile("mecab-dict-", ".tar.gz");
-    try {
-      DownloadUtil.download(archive, downloaded, expectedSha512);
-      try (InputStream in = Files.newInputStream(downloaded)) {
-        return extract(in, targetDirectory);
-      }
-    } finally {
-      Files.deleteIfExists(downloaded);
+    try (InputStream in = Files.newInputStream(Path.of(archive))) {
+      return extract(in, targetDirectory);
     }
-  }
-
-  /**
-   * Downloads a dictionary named in {@link DictionaryCatalog} and unpacks it. Requires
-   * {@code -Dopennlp.download.remote=true}.
-   *
-   * @param dictionaryId The catalog id, for example {@code mecab.ipadic} or
-   *                     {@code mecab.ko-dic}. Must not be {@code null}.
-   * @param targetDirectory The directory to unpack into; created when absent. Must not
-   *                        be {@code null}.
-   * @return The number of dictionary files extracted.
-   * @throws IOException Thrown if the catalog entry is missing, remote downloads are
-   *         disabled, or install fails.
-   * @throws IllegalArgumentException Thrown if a parameter is {@code null}.
-   */
-  public static int installFromCatalog(String dictionaryId, Path targetDirectory)
-      throws IOException {
-    if (dictionaryId == null) {
-      throw new IllegalArgumentException("dictionaryId must not be null");
-    }
-    if (targetDirectory == null) {
-      throw new IllegalArgumentException("targetDirectory must not be null");
-    }
-    final Path downloaded = Files.createTempFile("mecab-dict-", ".tar.gz");
-    try {
-      DictionaryCatalog.loadDefault().download(dictionaryId, downloaded);
-      return install(downloaded.toUri(), targetDirectory);
-    } finally {
-      Files.deleteIfExists(downloaded);
-    }
-  }
-
-  /**
-   * {@return {@code true} when {@code archive} uses the {@code file} scheme}
-   *
-   * @param archive The absolute archive URI.
-   */
-  private static boolean isLocalFile(URI archive) {
-    return "file".equalsIgnoreCase(archive.getScheme());
   }
 
   /**
@@ -245,10 +170,10 @@ public final class MecabDictionaryInstaller {
    *                      {@code null}. Not closed.
    * @param targetDirectory The directory to unpack into; created when absent. Must not
    *                        be {@code null}.
-   * @param maxEntryBytes Inclusive ceiling on one entry's declared size.
-   * @param maxTotalBytes Inclusive ceiling on total extracted bytes.
-   * @param maxEntries Inclusive ceiling on extracted dictionary file count.
-   * @param maxGzipRatio Inclusive ceiling on decompressed bytes per compressed byte.
+   * @param maxEntryBytes Inclusive limit on one entry's declared size.
+   * @param maxTotalBytes Inclusive limit on total extracted bytes.
+   * @param maxEntries Inclusive limit on extracted dictionary file count.
+   * @param maxGzipRatio Inclusive limit on decompressed bytes per compressed byte.
    * @return The number of dictionary files extracted.
    * @throws IOException Thrown if reading or writing fails, the archive contains no
    *         dictionary file, or a budget is exceeded.
@@ -263,6 +188,55 @@ public final class MecabDictionaryInstaller {
       throw new IllegalArgumentException("targetDirectory must not be null");
     }
     Files.createDirectories(targetDirectory);
+    final Path stagingDirectory = Files.createTempDirectory(targetDirectory, STAGING_PREFIX);
+    final List<Path> stagedFiles = new ArrayList<>();
+    final List<Path> publishedFiles = new ArrayList<>();
+    boolean published = false;
+    try {
+      final int extracted = extractToStaging(archiveStream, stagingDirectory, stagedFiles,
+          maxEntryBytes, maxTotalBytes, maxEntries, maxGzipRatio);
+      for (final Path stagedFile : stagedFiles) {
+        final Path target = targetDirectory.resolve(stagedFile.getFileName());
+        if (Files.exists(target)) {
+          throw new IOException("dictionary file already exists: " + target);
+        }
+      }
+      for (final Path stagedFile : stagedFiles) {
+        final Path target = targetDirectory.resolve(stagedFile.getFileName());
+        Files.move(stagedFile, target);
+        publishedFiles.add(target);
+      }
+      published = true;
+      return extracted;
+    } finally {
+      if (!published) {
+        for (final Path file : publishedFiles) {
+          Files.deleteIfExists(file);
+        }
+      }
+      for (final Path file : stagedFiles) {
+        Files.deleteIfExists(file);
+      }
+      Files.deleteIfExists(stagingDirectory);
+    }
+  }
+
+  /**
+   * Validates and extracts an archive into a temporary directory.
+   *
+   * @param archiveStream The gzip-compressed archive. Not closed.
+   * @param stagingDirectory The empty directory that receives extracted files.
+   * @param stagedFiles Receives each extracted file.
+   * @param maxEntryBytes Inclusive limit on one entry's declared size.
+   * @param maxTotalBytes Inclusive limit on total extracted bytes.
+   * @param maxEntries Inclusive limit on extracted dictionary file count.
+   * @param maxGzipRatio Inclusive limit on decompressed bytes per compressed byte.
+   * @return The number of dictionary files extracted.
+   * @throws IOException Thrown if validation, reading, or writing fails.
+   */
+  private static int extractToStaging(InputStream archiveStream, Path stagingDirectory,
+      List<Path> stagedFiles, long maxEntryBytes, long maxTotalBytes, int maxEntries,
+      int maxGzipRatio) throws IOException {
     final CountingInputStream compressed = new CountingInputStream(archiveStream);
     try (GZIPInputStream gzip = new GZIPInputStream(
         new UncloseableInputStream(compressed))) {
@@ -275,6 +249,7 @@ public final class MecabDictionaryInstaller {
         if (isEndBlock(header)) {
           break;
         }
+        verifyHeaderChecksum(header);
         final String name = headerName(header);
         final long size = headerSize(header);
         if (size > maxEntryBytes) {
@@ -283,25 +258,30 @@ public final class MecabDictionaryInstaller {
         }
         final char type = (char) header[TAR_TYPE_OFFSET];
         final String baseName = baseName(name);
-        // Only the archive root holds dictionary payload. Deeper files such as
+        // Only the archive root contains dictionary payload. Deeper files such as
         // mecab-ko-dic's user-dic templates carry empty numeric fields for
         // mecab-dict-index and would fail the load, or on a case-insensitive file
         // system overwrite a real lexicon file of the same base name.
         final boolean wanted = (type == '0' || type == 0) && pathDepth(name) <= 2
-            && (baseName.endsWith(".csv") || baseName.endsWith(".def")
-                || "dicrc".equals(baseName));
+            && (baseName.endsWith(MecabDictionary.LEXICON_EXTENSION)
+                || baseName.endsWith(MecabDictionary.DEFINITION_EXTENSION)
+                || MecabDictionary.CONFIGURATION_FILE.equals(baseName));
         if (wanted) {
           if (extracted >= maxEntries) {
             throw new IOException(
                 "extracted entry count exceeds safe limit of " + maxEntries);
           }
-          if (totalExtracted + size > maxTotalBytes) {
+          if (size > maxTotalBytes || totalExtracted > maxTotalBytes - size) {
             throw new IOException(
                 "extracted archive size exceeds safe limit of " + maxTotalBytes);
           }
-          final Path file = targetDirectory.resolve(baseName);
+          final Path file = stagingDirectory.resolve(baseName);
+          if (Files.exists(file)) {
+            throw new IOException("duplicate dictionary file in archive: " + baseName);
+          }
+          stagedFiles.add(file);
           try (InputStream entry = boundedStream(tar, size)) {
-            Files.copy(entry, file, StandardCopyOption.REPLACE_EXISTING);
+            Files.copy(entry, file);
           }
           extracted++;
           totalExtracted += size;
@@ -360,7 +340,7 @@ public final class MecabDictionaryInstaller {
    * Reads the NUL-terminated entry name from a tar header block.
    *
    * @param header The header block.
-   * @return The entry name. Never {@code null}.
+   * @return The entry name. Not {@code null}.
    */
   private static String headerName(byte[] header) {
     int end = 0;
@@ -375,17 +355,51 @@ public final class MecabDictionaryInstaller {
    *
    * @param header The header block.
    * @return The entry size in bytes.
-   * @throws IOException Thrown if the size field holds a non-octal digit.
+   * @throws IOException Thrown if the size field contains a non-octal digit.
    */
   private static long headerSize(byte[] header) throws IOException {
+    return parseOctalField(header, TAR_SIZE_OFFSET, TAR_SIZE_LENGTH, "size");
+  }
+
+  /**
+   * Verifies the checksum stored in a tar header.
+   *
+   * @param header The header block.
+   * @throws IOException Thrown if the checksum is malformed or does not match.
+   */
+  private static void verifyHeaderChecksum(byte[] header) throws IOException {
+    final long expected = parseOctalField(
+        header, TAR_CHECKSUM_OFFSET, TAR_CHECKSUM_LENGTH, "checksum");
+    long actual = 0;
+    for (int i = 0; i < header.length; i++) {
+      actual += i >= TAR_CHECKSUM_OFFSET && i < TAR_CHECKSUM_OFFSET + TAR_CHECKSUM_LENGTH
+          ? TAR_CHECKSUM_SPACE : header[i] & 0xFF;
+    }
+    if (expected != actual) {
+      throw new IOException("tar header checksum does not match");
+    }
+  }
+
+  /**
+   * Reads an octal numeric field from a tar header.
+   *
+   * @param header The header block.
+   * @param offset The field offset.
+   * @param length The field length.
+   * @param name The field name used in an error message.
+   * @return The parsed value.
+   * @throws IOException Thrown if the field contains a non-octal digit.
+   */
+  private static long parseOctalField(byte[] header, int offset, int length, String name)
+      throws IOException {
     long size = 0;
-    for (int i = TAR_SIZE_OFFSET; i < TAR_SIZE_OFFSET + TAR_SIZE_LENGTH; i++) {
+    for (int i = offset; i < offset + length; i++) {
       final byte b = header[i];
       if (b == 0 || b == ' ') {
         continue;
       }
       if (b < '0' || b > '7') {
-        throw new IOException("malformed tar size field");
+        throw new IOException("malformed tar " + name + " field");
       }
       size = size * 8 + (b - '0');
     }
@@ -396,7 +410,7 @@ public final class MecabDictionaryInstaller {
    * Strips any directory prefix from an archive entry name.
    *
    * @param name The entry name as stored in the archive.
-   * @return The part after the last {@code /}, or the whole name when there is none.
+   * @return The part after the last {@code /}, or the complete name when there is none.
    */
   private static String baseName(String name) {
     final int slash = name.lastIndexOf('/');
@@ -413,16 +427,20 @@ public final class MecabDictionaryInstaller {
    */
   private static int pathDepth(String name) {
     int depth = 0;
-    for (String segment : name.split("/")) {
-      if (!segment.isEmpty() && !".".equals(segment)) {
-        depth++;
+    int start = 0;
+    for (int i = 0; i <= name.length(); i++) {
+      if (i == name.length() || name.charAt(i) == '/') {
+        if (i > start && !(i - start == 1 && name.charAt(start) == '.')) {
+          depth++;
+        }
+        start = i + 1;
       }
     }
     return depth;
   }
 
   /**
-   * Computes the padding after an entry: tar content is stored in whole blocks.
+   * Computes the padding after an entry: tar content is stored in complete blocks.
    *
    * @param size The entry size in bytes.
    * @return The number of padding bytes up to the next block boundary.
@@ -457,7 +475,7 @@ public final class MecabDictionaryInstaller {
    * @param in The tar stream, positioned at the entry's first byte.
    * @param size The entry size in bytes.
    * @return A stream reporting end of stream after that many bytes, and failing if the
-   *         tar stream ends first. Never {@code null}; closing it leaves {@code in}
+   *         tar stream ends first. Not {@code null}; closing it leaves {@code in}
    *         open and positioned after the entry content.
    */
   private static InputStream boundedStream(InputStream in, long size) {
