--- opennlp-core/opennlp-runtime/src/main/resources/opennlp/tools/util/dictionary-catalog.properties
@@ -1,57 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License. You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-# Pinned remote dictionary files. OpenNLP ships URLs and SHA-512 digests only;
-# the data itself is never bundled. Fetching requires -Dopennlp.download.remote=true.
-
-# MeCab IPADIC 2.7.0 (EUC-JP). Upstream: MeCab project on SourceForge.
-mecab.ipadic.url=https://downloads.sourceforge.net/project/mecab/\
-mecab-ipadic/2.7.0-20070801/mecab-ipadic-2.7.0-20070801.tar.gz
-mecab.ipadic.sha512=35ea662cb62f1967849f7ed5781bd6dafef0fe20d63e88d9\
-a0057666e57ed23d5a0e6fb8d0701a0cc4da43a1050c1b0246\
-3bb862decc71c36b7fc2acdc158d86
-mecab.ipadic.filename=mecab-ipadic-2.7.0-20070801.tar.gz
-
-# mecab-ko-dic 2.1.1 (UTF-8). Upstream: eunjeon/mecab-ko-dic on Bitbucket.
-mecab.ko-dic.url=https://bitbucket.org/eunjeon/mecab-ko-dic/downloads/\
-mecab-ko-dic-2.1.1-20180720.tar.gz
-mecab.ko-dic.sha512=986f8f9c66c53accd296756bf632c979d2d44b695ada33f3\
-6c662f210dba34cd95d67b61dd8c84a1f7d59f80ee6bc22eb1\
-e9afb5dc6a7f9b6b75b4fbf2f8164f
-mecab.ko-dic.filename=mecab-ko-dic-2.1.1-20180720.tar.gz
-
-# LibreOffice en_US Hunspell pair, pinned to dictionaries commit 208a9fd8.
-hunspell.en_US.aff.url=https://raw.githubusercontent.com/LibreOffice/\
-dictionaries/208a9fd80b2a182fe20f224cd615119c6323ae2e/en/en_US.aff
-hunspell.en_US.aff.sha512=2b4448dfdff03caf300914415f4642f8d2ba5b650c5f024a\
-12355b420a279ffc12146649fce092ba591504476634a3d6\
-fd4c079335a27085b396fa76bfd28b74
-hunspell.en_US.aff.filename=en_US.aff
-
-hunspell.en_US.dic.url=https://raw.githubusercontent.com/LibreOffice/\
-dictionaries/208a9fd80b2a182fe20f224cd615119c6323ae2e/en/en_US.dic
-hunspell.en_US.dic.sha512=4be737249a8a436d20a02be575dcf6cf2f06f5f2abb840ea\
-5ec0ef0ac73a71fa0e4669e527c703d5c6b50ef61713a674\
-1b55bc136d559a54bcdeebcd62027988
-hunspell.en_US.dic.filename=en_US.dic
-
-hunspell.en_US.readme.url=https://raw.githubusercontent.com/LibreOffice/\
-dictionaries/208a9fd80b2a182fe20f224cd615119c6323ae2e/en/README_en_US.txt
-hunspell.en_US.readme.sha512=aa23ebc8adc0649b540264c7bf98cef5b6e383fec0e4a1a7\
-dd49d1c887cfeefd8edf6a568afc8a651521a3864c5b1ab5\
-0748ad16230d386809080b5b09135082
-hunspell.en_US.readme.filename=README_en_US.txt
